devday/requirements.txt
cmsplugin-filer==1.1.2
dj-database-url==0.4.1
Django==1.9.8
django-appconf==1.0.2
django-classy-tags==0.7.2
django-cms==3.3.1
django-filer==1.2.4
django-formtools==1.0
django-jsonfield==1.0.1
django-mptt==0.8.5
django-polymorphic==0.9.2
django-reversion==1.10.2
django-sekizai==0.9.0
Django-Select2==5.8.6
django-treebeard==4.0.1
djangocms-admin-style==1.2.3
djangocms-attributes-field==0.1.1
djangocms-column==1.6.0
djangocms-file==1.0
djangocms-googlemap==0.5.1
djangocms-inherit==0.2.1
djangocms-link==1.8.2
djangocms-picture==1.0.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==3.0.1
djangocms-video==1.0.0
easy-thumbnails==2.3
html5lib==0.9999999
Pillow==3.3.0
psycopg2==2.6.2
pytz==2016.6.1
six==1.10.0
tzlocal==1.2.2
Unidecode==0.4.19
webencodings==0.5
django-registration==2.1.2
django-bootstrap-form==3.2.1